DI02_03: Load files from Component

+++ src/main/java/com/carbonell/melodyseer/models/MyFile.java
@@ -4,6 +4,8 @@
  */
 package com.carbonell.melodyseer.models;
 
+import com.carbonell.melody.seer.component.api.Media;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
@@ -19,11 +21,18 @@ import java.io.Serializable;
 public class MyFile implements Serializable{
     private File file;
     private LocalDate downloadDate;
+    @JsonIgnore
+    private Media media;
     
     public MyFile(File file)
     {
         this.file = file;
         downloadDate = LocalDate.now();
+    }
+    
+    public MyFile(Media media)
+    {
+        this.media = media;
     }
     
     
@@ -33,16 +42,19 @@ public class MyFile implements Serializable{
 }
 
     public String getFileName() {
+        if(file == null) return media.mediaFileName;
         return file.getName();
     }
 
 
     public long getSize() {
+        if(file == null) return 0;
         return file.length();
     }
 
 
     public String getMime() throws IOException {
+        if(file == null) return media.mediaMimeType;
         return Files.probeContentType(file.toPath());
     }
 
@@ -53,10 +65,12 @@ public class MyFile implements Serializable{
 
     public String getDownloadDate() {
         // https://jenkov.com/tutorials/java-internationalization/simpledateformat.html
+        if(downloadDate == null) return "";
         return downloadDate.format(DateTimeFormatter.ofPattern("dd-MM-yyyy"));
     } 
     
     public void delete() {
+        if(file == null) return;
         file.delete();
     }
